Support <script type=webbundle>, as an alternative of <link rel=webbundle>
The motivation came from the public-web-perf discussion [*1], where
we've heard a security concern about using a <link> element.

This CL supports <script type=webbundle>, as an alternative, because a
<script> element is already considered as *unsafe*.

The tentative syntax is:

<script type=webbundle>
    {
      "href": "../resources/wbn/subresource.wbn",
      "resources": ["http://web-platform.test:8001/root.js",
                    "http://web-platform.test:8001/submodule.js"]
    }
</script>

That is equivalent to:

 <link rel="webbundle" href="../resources/wbn/subresource.wbn"
        resources="http://web-platform.test:8001/root.js http://web-platform.test:8001/submodule.js" />

We don't have any explainer nor an document mentioning this syntax yet.

Once we are sure that <script type=webbundle> can be used as a
replacement of <link rel=webbundle>, we will update the explainer [*2]
and make its behavior clear.

Regarding the implementation, we probably drop the <link
rel=webbundle> support in follow-up CLs.

This CL adds a very minimum test for <script type=webbundle> to avoid
duplicated efforts. Before we drop <link rel=webbundle ...>, we will
rewrite the existing tests, replacing <link rel=webbundle> with
<script type=webbundle>. That would uncover an issue of <script
type=webbundle>.

Other considerations and future plans:

- No support for dynamic update: Once a script element is inserted
  into a document, any update on the element, such as
  element.textContent = JSON.stringify(..), should be no-op.
  We'll add more tests once we can figure out the desired behavior.

- Feature detection: Currently, there is no easy way.
  We might want to add script.typeList-ish API as link.relList does.

- [1] https://lists.w3.org/Archives/Public/public-web-perf/2020Aug/0043.html
- [2] https://github.com/WICG/webpackage/blob/master/explainers/subresource-loading.md

Bug: 1082020
Change-Id: I94bf68f6a08610b202f2b926eb2d4ed825f0198d