Merge ImageData::data_union into ImageData::data
The ImageData interface has two data entries: data(array of uint8) and
data_union(could be array of uint8, float16 or float32). In this cl, I
deleted data and rename data_union to data. This cl prepares canvas to
work with more data types, which will help us to enable canvas to work
with p3 and rec2020.

Note that some part of the code works with uint8 only, i have put a TODO
for each of the occurrence. I will address them in the next cl.

Bug: 1115317

Change-Id: Ia7f203e627547eceb6da479522b3130bc6eaef12